Rename swsh party menu file
